304: Rename some frontend files
