+++ frontend/src/pages/cans/detail/getCan.js
